fix file name problem

+++ delete_blog.php
@@ -1,6 +1,9 @@
 <?php 
-    $blog= $_POST['filename'];
-    unlink($blog);
-    header("Location: index.php");
+    foreach($_POST['filename'] as $blog) {
+        if ($blog) {
+            unlink($blog);
+            header("Location: index.php");
+        }
+    }
     exit();
 ?>
